+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - add default preset timer values for fresh install.
 
+## Fixed
+
+- Fixed the progress bar animation and its initial and final states.
+
 ## v1.7.0
 
 ### Improved
@@ -18,12 +22,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improved release notes with better Markdown formatting.
 - Improved the app update experience and interface.
 
-### Bug fixes
+### Fixed
 
 - Fixed the timer progress bar position when reopening the window.
 
 ## v1.6.1
 
-### Bug Fixes
+### Fixed
 
 - Fixed an issue where release notes were not displayed when updating to a new version.